Address compiler warnings (3/3)
- remove /config dir (`use Mix.Config` deprecated in Elixir v1.9)

`config.exs` is not being used for anything — and is no longer being
generated by `mix new` as of v1.9 — so deleted to avoid deprecation
warnings. This also avoids having to advance the minimum Elixir version
to 1.9 since changing `config.exs` to using `import Config` yields a
compiler *error* (not warning) for *any version* of Elixir prior to 1.9.